Computer_Vision/3_Mnist_Flask_Deployment/requirements.txt
flask == 1.0.2
tensorflow == 1.9.0
numpy == 1.15.4
imageio == 2.2.0
pillow == 4.2.1
